+++ RELEASES.md
@@ -1,5 +1,15 @@
 # Confidential Match Source Code Releases
 
+## September 18, 2025
+
+* Version: v1.260.0
+* Lookup Service
+    * Image Digest: sha256:f2f9d08ef228d1cdc1f9cdd0893d8cfdd24ac61e518f11d13c517c7b7e1efba5
+    * [Signature](https://us-docker.pkg.dev/admcloud-cfm-public/docker-repo-signatures/lookup_server_gcp_signature@sha256:c8d2c9c311959e7d0b853d3062852b84104c3f1b64e1d1fceb41e95fd0e6584d)
+* Match Request Processor
+    * Image Digest: sha256:dd336dc7da5a3d1c39ce73d20681c18f8ebf52f87f2ee5b3d2819b32574bd997
+    * [Signature](https://us-docker.pkg.dev/admcloud-cfm-public/docker-repo-signatures/mrp_app_gcp_signature@sha256:f002ad1ee0b0bb89a370069fac9bf5b2b92fef40fec6ddf38ce1e5c7388f9632)
+
 ## September 5, 2025
 
 * Version: v1.249.0